feat: maturity-mcp-host-guide-001 - Follow A Complete Runtime-Backed Factory Session Transcript
